+++ src/VehicleGroupStatistics.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
-import { Grid, Header, Input, Menu } from 'semantic-ui-react'
+import { Button, Grid, Header, Input, Menu } from 'semantic-ui-react'
 import VehicleGroup from './VehicleGroup'
 import VehicleGroupStatistic from './VehicleGroupStatistic'
 
@@ -10,6 +10,8 @@ import VehicleGroupStatistic from './VehicleGroupStatistic'
  * @param {VehicleGroup[]} props.vehicleGroup the vehicle group data to display
  */
 export default class VehicleGroupStatistics extends Component {
+  static defaultProps = { hidden: false, searchTerm: '' }
+
   static propTypes =
     { vehicleGroups: PropTypes.arrayOf(PropTypes.instanceOf(VehicleGroup)) }
 
@@ -21,7 +23,41 @@ export default class VehicleGroupStatistics extends Component {
 
   constructor(props) {
     super(props)
-    this.state = { searchTerm: '' }
+    this.state = { hidden: props.hidden, searchTerm: props.searchTerm }
+  }
+
+  /**
+   * @return {boolean}
+   *  if the component is "hidden"
+   */
+  hidden() {
+    if (this.state && typeof this.state.hidden !== 'undefined') {
+      return this.state.hidden
+    }
+    return false
+  }
+
+  /**
+   * @return {string}
+   *  the "Hide" button's visible text
+   */
+  get hideButtonText() {
+    if (this.hidden()) {
+      return 'Show'
+    } else {
+      return 'Hide'
+    }
+  }
+
+  /**
+   * Needs to be binded to `this`.
+   * @see {@link https://reactjs.org/docs/handling-events.html}
+   * @method
+   * @param {SyntheticEvent} event
+   *  https://reactjs.org/docs/events.html
+   */
+  onHideButtonClick = (event) => {
+    this.setState({ hidden: !this.state.hidden })
   }
 
   /**
@@ -39,8 +75,12 @@ export default class VehicleGroupStatistics extends Component {
    * {@link https://reactjs.org/docs/react-component.html#render}
    */
   render() {
+    const hidden = this.hidden()
+    const hideButtonText = this.hideButtonText
+    const onHideButtonClick = this.onHideButtonClick
     const onSearchChange = this.onSearchChange
     const vehicleGroups = this.vehicleGroups
+    const visible = !hidden
     return(
       <Grid.Row color='black'>
         <Grid.Column width={16}>
@@ -48,15 +88,23 @@ export default class VehicleGroupStatistics extends Component {
             <Menu.Item header>
               <Header as='h2' color='blue' content='Vehicle Info.' inverted/>
             </Menu.Item>
+            <Menu.Item>
+              <Button.Group compact inverted toggle>
+                <Button active={visible} basic content={hideButtonText}
+                  color='black' inverted onClick={onHideButtonClick}/>
+              </Button.Group>
+            </Menu.Item>
+            {visible &&
             <Menu.Menu position='right'>
               <Menu.Item>
                 <Input icon='search' onChange={onSearchChange}
                   placeholder='Search...'/>
               </Menu.Item>
             </Menu.Menu>
+            }
           </Menu>
         </Grid.Column>
-        {vehicleGroups.map((vehicleGroup) =>
+        {visible && vehicleGroups.map((vehicleGroup) =>
         <Grid.Column computer={5} key={vehicleGroup.key} largeScreen={4}
           tablet={6}>
           <VehicleGroupStatistic vehicleGroup={vehicleGroup}/>